Don't bail update when encountering UID, PHOTO or an empty value
The break statement broke out of the loop completely. Due to this behavior only fields which were defined before one of the mentioned fields was update.

In my case this meant that VERSION would be updated, but UID was the second field and the rest wouldn't update after selected 'Pull' from the dropdown menu.